Correctly handle string encodings in java-readline
Previously java-readline always assumed either a C or latin1 locale.
Using java-readline with a UTF-8 locale with non-ascii characters would
result in mangled input and output for characters outside the ascii
range.

This commit reworks the string handling to respect the current locale
and to also work with characters outside the BMP when in a UTF-8 locale.

This might be possible to fix by using native C-code, but it seemed
simpler to just use JNI to have java use the default encoding for
converting the bytes between C and JNI.

Calling a readline function will return an array of bytes in the default
platform encoding, which I now convert to a jstring by calling:

new String(byte[])

from JNI, which will use the default platform encoding. The opposite is
also true. Given a jstring, I convert to a char* by calling:

string.getBytes()

from JNI, which will also use the default platform encoding.

The bats test suite for this is now passing for all cases.