Update maven build for JDKs > 1.8.
Tweak Makefile for detecting older JDKs.

Most of this is probably not really needed, but if you are building with
JDK 10 or newer, we will set source and target to JDK8. Otherwise, if
building with 1.8 or greater, we set the source and target to 1.8.
Otherwise, they are set to 1.5. This keeps the settings as they were
before.