Cast const char** in globalStringInternals to char**.
Newer compilers error on this. They should do that - the code is
incorrect - we are putting const char** values into this global array.
If someone uses setVarStringImpl to write to one of these, it will
crash.

This is however already mentioned in the documentation:

https://github.com/aclemons/java-readline/blob/99fe57e6c3544be476187fb7c66a664e6326c3ea/src/native/org_gnu_readline_Readline.c#L63-L64

and they all had a comment in front indicating they were:

```c
const `/* const */*
```

Basically we are relying on code using java-readline to not call the
wrong thing.

It has been this way for more than 20 years, so for now I will not
change this and will simply cast away the compiler error.

Related to #41